[WIP] statefulness of Chat
